Data.OpenADT

Description

This module defines the OpenADT type, which is an algebraic data type with constructors defined by its argument's row type.

Documentation

type OpenADT r = Fix (VarF r)

A algebraic data type that can have constructors added and removed.

module Data.OpenADT.TH

module Data.OpenADT.VarF

module Data.OpenADT.VariantsF